+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.0.0
+-----
+
+* Fixed issue with too big terms in case of recursive types
+* Recpect the `size` argument
+* Types with parameters require `Arg` now
+
 0.2.2
 -----